Getting a clean copy for a merge ... these commits are untested and could/should be blown away.
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/coamps@11810 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 2e7b4dc9bf1b85cd1199ccc7b0f384dc5b85d13e

+++ assimilation_code/location/threed_sphere/location_mod.f90
@@ -1396,7 +1396,7 @@ subroutine get_close_obs(gc, base_loc, base_type, locs, loc_qtys, loc_types, &
 type(get_close_type),          intent(in)  :: gc
 type(location_type),           intent(inout) :: base_loc, locs(:)
 integer,                       intent(in)  :: base_type, loc_qtys(:), loc_types(:)
-integer,                       intent(out) :: num_close, close_ind(:)
+integer(i8),                   intent(out) :: num_close, close_ind(:)
 real(r8),            optional, intent(out) :: dist(:)
 type(ensemble_type), optional, intent(in)  :: ens_handle
 
@@ -1418,7 +1418,7 @@ type(get_close_type),          intent(in)  :: gc
 type(location_type),           intent(inout)  :: base_loc, locs(:)
 integer,                       intent(in)  :: base_type, loc_qtys(:)
 integer(i8),                   intent(in)  :: loc_indx(:)
-integer,                       intent(out) :: num_close, close_ind(:)
+integer(i8),                   intent(out) :: num_close, close_ind(:)
 real(r8),            optional, intent(out) :: dist(:)
 type(ensemble_type), optional, intent(in)  :: ens_handle
 
@@ -1439,7 +1439,7 @@ subroutine get_close(gc, base_loc, base_type, locs, loc_qtys, &
 type(get_close_type),          intent(in)  :: gc
 type(location_type),           intent(inout)  :: base_loc, locs(:)
 integer,                       intent(in)  :: base_type, loc_qtys(:)
-integer,                       intent(out) :: num_close, close_ind(:)
+integer(i8),                   intent(out) :: num_close, close_ind(:)
 real(r8),            optional, intent(out) :: dist(:)
 type(ensemble_type), optional, intent(in)  :: ens_handle
